Extract calculation logic into seperate classes
