+++ src/instruments/src/EFB/Performance/Calculators/CommonCalculations.ts
